# internal/server/docs/swagger.yaml
definitions:
  config.AppInfo:
    properties:
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
    type: object
  server.AllowedAppsResponse:
    properties:
      apps:
        items:
          $ref: '#/definitions/config.AppInfo'
        type: array
    type: object
  server.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  server.OpenInRequest:
    properties:
      app:
        description: App ID (e.g., "finder", "vscode", "cursor")
        type: string
      path:
        description: Path to open
        type: string
    type: object
  server.OpenInResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  server.ProjectsResponse:
    properties:
      projects:
        items:
          $ref: '#/definitions/thinkt.Project'
        type: array
    type: object
...
  server.SessionsResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/thinkt.SessionMeta'
        type: array
    type: object
  server.SourceInfo:
    properties:
      available:
        type: boolean
      base_path:
        type: string
      name:
        type: string
    type: object
  server.SourcesResponse:
    properties:
      sources:
        items:
          $ref: '#/definitions/server.SourceInfo'
        type: array
    type: object
  server.TeamMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/thinkt.TeamMessage'
        type: array
    type: object
  server.TeamTasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/thinkt.TeamTask'
        type: array
    type: object
  server.TeamsResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/thinkt.Team'
        type: array
    type: object
  server.ThemeColors:
    properties:
      accent:
        description: UI chrome
        type: string
      assistant_block:
        $ref: '#/definitions/server.ThemeStyle'
      assistant_label:
        $ref: '#/definitions/server.ThemeStyle'
      border_active:
        type: string
      border_inactive:
        type: string
      confirm_prompt:
        allOf:
        - $ref: '#/definitions/server.ThemeStyle'
        description: Confirm dialog
      confirm_selected:
        $ref: '#/definitions/server.ThemeStyle'
      confirm_unselected:
        $ref: '#/definitions/server.ThemeStyle'
      text_muted:
        $ref: '#/definitions/server.ThemeStyle'
      text_primary:
        allOf:
        - $ref: '#/definitions/server.ThemeStyle'
        description: Text styles
      text_secondary:
        $ref: '#/definitions/server.ThemeStyle'
      thinking_block:
        $ref: '#/definitions/server.ThemeStyle'
      thinking_label:
        $ref: '#/definitions/server.ThemeStyle'
      tool_call_block:
        $ref: '#/definitions/server.ThemeStyle'
      tool_label:
        $ref: '#/definitions/server.ThemeStyle'
      tool_result_block:
        $ref: '#/definitions/server.ThemeStyle'
      user_block:
        allOf:
        - $ref: '#/definitions/server.ThemeStyle'
        description: Conversation blocks
      user_label:
        allOf:
        - $ref: '#/definitions/server.ThemeStyle'
        description: Labels
    type: object
  server.ThemeInfo:
    properties:
      active:
        type: boolean
      colors:
        $ref: '#/definitions/server.ThemeColors'
      description:
        type: string
      embedded:
        type: boolean
      name:
        type: string
    type: object
  server.ThemeStyle:
    properties:
      bg:
        type: string
      bold:
        type: boolean
      fg:
        type: string
      italic:
        type: boolean
      underline:
        type: boolean
    type: object
  server.ThemesResponse:
    properties:
      active:
        type: string
      themes:
        items:
          $ref: '#/definitions/server.ThemeInfo'
        type: array
    type: object
...
  thinkt.Project:
    properties:
      display_path:
        description: Human-readable path
        type: string
      id:
        type: string
      last_modified:
        type: string
      name:
        description: Display name
        type: string
      path:
        description: Full filesystem path
        type: string
      path_exists:
        description: Whether the project directory still exists on disk
        type: boolean
      session_count:
        type: integer
      source:
        allOf:
        - $ref: '#/definitions/thinkt.Source'
        description: Which tool
      source_base_path:
        description: Root storage path for this source
        type: string
      workspace_id:
        description: Which machine/host
        type: string
    type: object
...
  thinkt.SessionMeta:
    properties:
      chunk_count:
        description: 'Number of files: 0=unknown, 1=single, 2+=chunked'
        type: integer
      created_at:
        type: string
      entry_count:
        type: integer
      file_size:
        description: Size in bytes
        type: integer
      first_prompt:
        type: string
      full_path:
        description: Path to session file
        type: string
      git_branch:
        type: string
      id:
        type: string
      model:
        type: string
      modified_at:
        type: string
      project_path:
        description: Normalized project path
        type: string
      source:
        allOf:
        - $ref: '#/definitions/thinkt.Source'
        description: Which tool (kimi, claude)
      summary:
        type: string
      workspace_id:
        description: Which machine/host
        type: string
    type: object
  thinkt.Source:
    enum:
    - thinkt
    - claude
    - copilot
    - gemini
    - kimi
    type: string
    x-enum-varnames:
    - SourceThinkt
    - SourceClaude
    - SourceCopilot
    - SourceGemini
    - SourceKimi
  thinkt.Team:
    properties:
      created_at:
        type: string
      description:
        type: string
      lead_agent_id:
        description: LeadAgentID is the config-format agent ID (e.g., "team-lead@myteam").
        type: string
      lead_session_id:
        description: |-
          LeadSessionID is the UUID of the team lead's session.
          Subagent sessions live under projects/{projectDir}/{LeadSessionID}/subagents/
        type: string
      members:
        items:
          $ref: '#/definitions/thinkt.TeamMember'
        type: array
      name:
        description: Name is the team identifier (directory name under ~/.claude/teams/).
        type: string
      source:
        allOf:
        - $ref: '#/definitions/thinkt.Source'
        description: Source provenance
      status:
        $ref: '#/definitions/thinkt.TeamStatus'
      workspace_id:
        type: string
    type: object
  thinkt.TeamMember:
    properties:
      agent_id:
        description: AgentID is the config-format ID (e.g., "researcher@myteam").
        type: string
      agent_type:
        description: AgentType classifies the member (e.g., "team-lead", "general-purpose").
        type: string
      color:
        type: string
      cwd:
        type: string
      joined_at:
        type: string
      model:
        type: string
      name:
        description: Name is the short display name (e.g., "researcher").
        type: string
      session_path:
        description: |-
          SessionPath is the path to this member's subagent JSONL file.
          Populated during discovery. Empty for unresolved members.
        type: string
      source_agent_id:
        description: |-
          SourceAgentID is the short hash used in session JSONL entries
          (e.g., "ab17e07"). Populated when correlating team config with
          session data. May be empty if not yet resolved.
        type: string
    type: object
  thinkt.TeamMessage:
    properties:
      color:
        type: string
      from:
        type: string
      read:
        type: boolean
      text:
        type: string
      timestamp:
        type: string
    type: object
  thinkt.TeamStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - TeamStatusActive
    - TeamStatusInactive
  thinkt.TeamTask:
    properties:
      active_form:
        type: string
      blocked_by:
        items:
          type: string
        type: array
      blocks:
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: string
      is_internal:
        type: boolean
      owner:
        type: string
      status:
        description: '"pending", "in_progress", "completed", "deleted"'
        type: string
      subject:
        type: string
    type: object
...
info:
  contact: {}
  description: |
    API for exploring AI conversation traces from Claude Code, Kimi Code, and other sources.

    ## Authentication

    The API supports Bearer token authentication. When `THINKT_API_TOKEN` environment variable
    is set or `--token` flag is provided, all requests must include the token in the
    Authorization header:

    Authorization: Bearer <token>

    Generate a secure token with: `thinkt serve token`
  title: Thinkt API
  version: "1.0"
paths:
  /open-in:
    post:
      consumes:
      - application/json
      description: Opens the specified path in an allowed application (e.g., Finder,
        VS Code)
      parameters:
      - description: Open-in request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.OpenInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.OpenInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Open a path in an application
      tags:
      - open-in
  /open-in/apps:
    get:
      description: Returns the list of enabled applications that can be used with
        open-in
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AllowedAppsResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List allowed apps for open-in
      tags:
      - open-in
  /projects:
    get:
      description: Returns all projects from all sources, optionally filtered by source
      parameters:
      - description: Filter by source (e.g., 'claude', 'kimi')
        in: query
        name: source
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ProjectsResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all projects
      tags:
      - projects
  /projects/{projectID}/sessions:
    get:
      description: Returns all sessions belonging to a specific project
      parameters:
      - description: Project ID (URL-encoded path)
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SessionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List sessions for a project
      tags:
      - sessions
...
  /sources:
    get:
      description: Returns all configured trace sources (e.g., Claude Code, Kimi Code)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SourcesResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List available trace sources
      tags:
      - sources
  /teams:
    get:
      description: Returns all Claude Code teams discovered on this machine
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TeamsResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all teams
      tags:
      - teams
  /teams/{teamName}:
    get:
      description: Returns team details including resolved member-to-session mappings
      parameters:
      - description: Team name
        in: path
        name: teamName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/thinkt.Team'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get team details
      tags:
      - teams
  /teams/{teamName}/members/{memberName}/messages:
    get:
      description: Returns inbox messages for a specific team member
      parameters:
      - description: Team name
        in: path
        name: teamName
        required: true
        type: string
      - description: Member name
        in: path
        name: memberName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TeamMessagesResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List team member messages
      tags:
      - teams
  /teams/{teamName}/tasks:
    get:
      description: Returns the shared task board for a team
      parameters:
      - description: Team name
        in: path
        name: teamName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TeamTasksResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List team tasks
      tags:
      - teams
  /themes:
    get:
      description: Returns all available themes (built-in and user themes) with their
        color definitions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ThemesResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List available themes
      tags:
      - themes
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"